+++ HISTORY.rst
@@ -2,6 +2,22 @@
 History
 =======
 
+0.4.0 (2018-03-16)
+------------------
+
+* Add Chocolate chip variation
+* Add long instructions option
+
+0.3.0 - 0.3.4 (2018-03-14)
+--------------------------
+
+* Final Base Version with CLI
+
+0.2.0 (2018-03-11)
+------------------
+
+* Add Docs
+
 0.1.0 (2018-03-10)
 ------------------